app/requirements.txt
flask==0.12
simplejson==3.10.0
numpy
scikit-surprise